Ejercicios parte 2.1 clase 3

+++ Modulo_1/6.1_Ejercicio_par_impar.py
@@ -1,5 +1,5 @@
 # Solicitar entradas
-num = float(input('Introduce el número: '))
+num = int(input('Introduce el número: '))
 
 # Código
 if num % 2 == 0:
